pyproject.toml
[project]
name = "audio-ecology"
version = "0.1.0"
description = "Passive acoustic monitoring pipeline"
authors = [{name = "TJGWilson"}]
requires-python = ">=3.11"

dependencies = [
    "polars",
    "pydantic>=2",
    "pyyaml",
    "typer",
]

[project.optional-dependencies]
birds = [
    "birdnet>=0.2.13",
]

[tool.ruff]
line-length = 88

[project.scripts]
audio-ecology = "audio_ecology.cli:app"
